+++ HISTORY.rst
@@ -2,6 +2,10 @@
 History
 =======
 
+* Add Django 1.9+ fixer to rename the legacy engine name ``django.db.backends.postgresql_psycopg2`` in ``settings.DATABASES`` to ``django.db.backends.postgresql``.
+
+  Thanks to Thibaut Decombe in `PR #280 <https://github.com/adamchainz/django-upgrade/pull/280>`__.
+
 1.11.0 (2022-10-26)
 -------------------